Make both input and output contigeous
