+++ setup.py
@@ -9,7 +9,7 @@ INSTALL_REQUIRES = [
     "numpy>=1.24.4", "scipy>=1.10.0", "scikit-image>=0.21.0",
     "PyYAML", "typing-extensions>=4.9.0",
     "pydantic>=2.7.0",
-    "albucore>=0.0.11",
+    "albucore>=0.0.13",
     "eval-type-backport"
 ]
