+++ tests/test_augmentations.py
@@ -1092,92 +1092,3 @@ def test_pixel_domain_adaptation(kind: str) -> None:
             atol=2 if img.dtype == np.uint8 else 0.01,
             err_msg=f"{adapted.mean()} {img.mean()} {ref_img.mean()}",
         )
-
-
-@pytest.mark.parametrize(
-    ["augmentation_cls", "params"],
-    get_transforms(
-        custom_arguments={
-            # only image
-            A.HistogramMatching: {
-                "reference_images": [SQUARE_UINT8_IMAGE],
-                "read_fn": lambda x: x,
-            },
-            A.FDA: {
-                "reference_images": [SQUARE_FLOAT_IMAGE],
-                "read_fn": lambda x: x,
-            },
-            A.PixelDistributionAdaptation: {
-                "reference_images": [SQUARE_FLOAT_IMAGE],
-                "read_fn": lambda x: x,
-                "transform_type": "standard",
-            },
-            A.MedianBlur: {"blur_limit": (3, 5)},
-            A.TemplateTransform: {
-                "templates": SQUARE_UINT8_IMAGE,
-            },
-            A.RingingOvershoot: {"blur_limit": (3, 5)},
-            # dual
-            A.Crop: {"y_min": 0, "y_max": 10, "x_min": 0, "x_max": 10},
-            A.CenterCrop: {"height": 10, "width": 10},
-            A.CropNonEmptyMaskIfExists: {"height": 10, "width": 10},
-            A.RandomCrop: {"height": 10, "width": 10},
-            A.RandomResizedCrop: {"height": 10, "width": 10},
-            A.RandomSizedCrop: {"min_max_height": (4, 8), "height": 10, "width": 10},
-            A.CropAndPad: {"px": 10},
-            A.Resize: {"height": 10, "width": 10},
-            A.RandomSizedBBoxSafeCrop: {"height": 10, "width": 10},
-            A.BBoxSafeRandomCrop: {"erosion_rate": 0.5},
-            A.XYMasking: {
-                "num_masks_x": (1, 3),
-                "num_masks_y": (1, 3),
-                "mask_x_length": 10,
-                "mask_y_length": 10,
-                "mask_fill_value": 1,
-                "fill_value": 0,
-            },
-             A.MixUp: {
-                "reference_data": [{"image": SQUARE_UINT8_IMAGE,
-                                    "mask": np.random.randint(0, 1, [100, 100, 3], dtype=np.uint8),
-                                    }],
-                "read_fn": lambda x: x,
-            },
-            A.TextImage: dict(font_path="./tests/files/LiberationSerif-Bold.ttf")
-        },
-    ),
-)
-def test_non_contiguous_input(augmentation_cls, params, bboxes):
-    image = np.empty([3, 100, 100], dtype=np.uint8).transpose(1, 2, 0)
-    mask = np.empty([3, 100, 100], dtype=np.uint8).transpose(1, 2, 0)
-
-    # check preconditions
-    assert not image.flags["C_CONTIGUOUS"]
-    assert not mask.flags["C_CONTIGUOUS"]
-
-    if augmentation_cls == A.RandomCropNearBBox:
-        # requires "cropping_bbox" arg
-        aug = augmentation_cls(p=1, **params)
-        aug(image=image, mask=mask, cropping_bbox=bboxes[0])
-    elif augmentation_cls in [A.RandomSizedBBoxSafeCrop, A.BBoxSafeRandomCrop]:
-        # requires "bboxes" arg
-        aug = A.Compose([augmentation_cls(p=1, **params)], bbox_params=A.BboxParams(format="pascal_voc"))
-        aug(image=image, mask=mask, bboxes=bboxes)
-    elif augmentation_cls == A.TextImage:
-        aug = A.Compose([augmentation_cls(p=1, **params)], bbox_params=A.BboxParams(format="pascal_voc"))
-        aug(image=image, mask=mask, bboxes=bboxes, textimage_metadata=[])
-    elif augmentation_cls == A.OverlayElements:
-        # requires "metadata" arg
-        aug = A.Compose([augmentation_cls(p=1, **params)])
-        aug(image=image, overlay_metadata=[], mask=mask)
-    else:
-        # standard args: image and mask
-        if augmentation_cls == A.FromFloat:
-            # requires float image
-            image = (image / 255).astype(np.float32)
-            assert not image.flags["C_CONTIGUOUS"]
-        elif augmentation_cls == A.MaskDropout:
-            # requires single channel mask
-            mask = mask[:, :, 0]
-
-        aug = augmentation_cls(p=1, **params)
-        aug(image=image, mask=mask)
